Refactor code, merge into SafariWebAuth
